Fall back to hex value when we can't look up a character
We're dependant on Perl for Unicode tables, so if it doesn't have a
recent enough set of definitions (e.g. Debian Stretch only has up to
Unicode 8) we were getting uninitialized value warnings for Unicode 9/10
characters. Fall back to outputting the hex value for the character.
Reported by Lars Wirzenius.